fix: move preprocessing logic to separate file
